Adds Mesh class
