+++ frontend/web/src/app/stl-viewer/stl-viewer.component.ts
@@ -25,11 +25,11 @@ export class StlViewerComponent implements OnInit {
 
 
     //========================================================================
-    var stlLoader = new StlLoader(null, null, null, null);
+/*    var stlLoader = new StlLoader(null, null, null, null);
     stlLoader.setDecimalPrecision(3);
     var loaderSelector = new LoaderSelector();
     loaderSelector.registerLoader('stl', stlLoader);
-
+*/
 
     //=========================================================================
     const viewer = new Viewer(canvas, parameters);
